requirements.txt
langchain
langchain-core
langchain-chroma
langchain-google-genai
langchain-groq
chromadb
python-dotenv
streamlit